requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.21.0
plotly>=5.20.0
pyyaml>=6.0.1
